Move to use Flask for the server functions
The current server, SimpleHTTPServer, provides for sufficient static
serving of files but lacks ability to do anything more complicated.
Switching to use Flask allows the creation of API endpoints to provide
the required level of system control.

In migrating to use Flask, some refactoring and reorganising of files
has been done to simplify.