Install npm using script when building package
The server struggles to install the npm package so use a script to
provide the npm build dependency. In the process NodeJS has to be
compiled so the build does take a while.

Potentially, either npm (and all of its dependencies) can be added to
our repository or, since NodeJS and npm are being built anyway without
the need for most of the dependencies which come from using the npm
package, NodeJS and npm could be built and packaged from the server,
eliminating the lengthy step of compiling it.