Add challenge description to the challenge menu
KanoComputing/peldins#1473
Provides details about the last challenge attempted and includes a quick
way to launch into it. Tracks the last challenge attempted and stores it
in localstorage. Adds descriptions to all of the current challenges.
Changes the styling on the challenge menu to bring it closer to that of
the Kano Blocks levels.