Add medium wallpaper size and scale wallpapers
KanoComputing/peldins#1487
Renames the 4:3 mode to 'medium' and adds a new 4:3 resolution
(1600x1200) to bring into line with the wallpapers already used in Kano
Desktop. Also scales the images when producing them; this isn't an ideal
solution as it is upscaled and we should probably alter the canvas to be
a higher resolution so we end up downscaling but this solution is
satisfactory for now.