Expand spellbook functionality and add 'for' spell
Adds the ability to add examples of the spells. Adds the 'for' spell and
example. Needs some refining to allow spells with different syntax to
the standard command arg1, arg2, ... (e.g. the for command) to be
displayed properly. The examples need better styling and examples need
to be added for most of the spells.