Show exec-time errors in ace
This commit catches execution errors of the resulting JS code, maps them
back to the original CoffeeScript code and generates annotations for ace
to mark the line where the error comes from.

To be able to do so, I had to add a new directive that serves as an API
between the display and editor ones. This directive feeds the annotations
to ace on top of the original ones which cover mostly coffeescript syntax.

So we now get both. Yay!