Move split dialogs to be styled centrally
Currently whenever there is a split dialog, the padding to the top half
to account for the bottom half must be set per dialog. This patch moves
the styling for this to a CSS class which can be applied.

Note: This works for the standard bottom of one row of plain buttons, if
there are more complicated bottom styles then dialog-specific styles
will need to be implemented.