Split saved data into code (.draw) and json files
The convention for Kano Share is to have the shared code in a file with
its metadata in a json and the preview picture in a png (all sharing the
same filename). The currently implemented system in Kano Draw is to just
save the picture and all the data (code and metadata) in a json with the
.draw extension. Split this .draw file into 2 - a .draw file containing
only the code and a .json file containing just the metadata.