Add app close button
Kano Draw will have no decoration when run on Kano OS so a way to close
it is needed. The request to close is handled by the server but this is
in a daemon process struggles to communicate to the parent process. To
combat this, send the parent PID to the child to be able to signal the
close event.