Fix production variable not being defined
The injection of the production variable to the window from the node
environment happens after the code that depends on it is loaded. This
means that no matter whether the variable is set, the app will remain in
non-production mode. Swap the order of scripts to allow this variable to
function correctly.