Fix close button when launched from desktop icon
When Kano Draw is launched from the desktop icon (and the entry in Apps)
the close button does not function correctly. The original method was to
send an interrupt signal to the Kano Draw process (from the server) to
signal shutdown. This fails so instead catch the interrupt signal and
handle it to clean up.