Add launching a save from the CLI
KanoComputing/peldins#1511
KanoComputing/peldins#1545
This allows launching of Kano Draw and opening a save at the same time
by using `kano-draw /path/to/save`. This has ramifications for both
launching shares from the Kano Profile GUI and also behaviour when
double clicking on a `.draw` file.

This implementation redirects the webapp to `/localLoad/path/to/save`
which in turn loads the file from the local server and redirects to the
playground in order to be able to view it. The behaviour is similar to
when launching an online share directly from the URL but because this
online functionality is already deployed and a slightly different case,
this solution does not integrate with it.