Add application/x-kano-draw MIME information
KanoComputing/peldins#1511
By adding this XML to `/usr/share/mime/packages`, the MIME type
application/x-kano-draw becomes associated with the .draw file
extension. This allows all files with the extension to be launched, by
clicking them, in Kano Draw. It also allows setting of the icon to be a
custom Kano Draw one by creating the icon file in the icon theme (as
`mimetypes/application-x-kano-draw.png`).