Wallpaper: Ensure that wallpapers have backgrounds
KanoComputing/peldins#1787
Currently, when exporting a wallpaper, the alpha from the created image
is retained which means that there is no background. This is a problem
because the canvas upon which the drawing is designed has a white
background, but when used as a wallpaper, the default background is
black. Fix this by artificially drawing a white background when exporting