Summer Camp Feedback: Use multiple question response endpoint
The API has been modified to allow answers to multiple questions to be
sent with one request so migrate to use this system.

This API change also removes the ability to specify the question text in
the response so change the request parameters.

TODO: Fill in actual Question IDs as soon as they are uploaded to the
database