Offline: Don't display the shares view if offline
When running on the Kano kit, HTTP requests are made to retrieve shares
for the gallery; these requests are made to the local server and, as no
endpoint has been added to actually perform these requests and, even if
it had, it would likely be offline anyway, these requests fail. Prevent
these errors from occurring by not showing the gallery when offline.