Offline: Prevent error on exit
KanoComputing/os-dashboard#90
An error message appears when attempting to close the app in offline
mode. This is because the server, instead of responding, just shuts down
and hence doesn't reply. This means that the application itself doesn't
get a response so assumes that something is wrong. Resolve this by
replying and not displaying a success message when this happens.