Remove trackEvent listener
The only event that seems to have been firing is the sdk, which fires a signedInToKanoWorld event on successful sign in. This was causing duplication with the in-app event. Since all other tracking is handled by the app, and no other events are triggering this, it seems best to remove the listener to prevent any other rogue events firing.

Trello card: https://trello.com/c/rjMJSqys/1621-2-implement-kano-code-tracking-on-make-art