Update login methods, to simply save the user and token that have already been returned
Previously, the app was duplicating the login methods – users would be logged in by the form, and then logged in again when the `core auth.login` method was called. This commit introduces a new `saveLogin` method to `api.auth` to save the token that has already been acquired to `localStorage` and to `xStorage`. This seems to be more reliable, saves the sessionacross domains, and seems to fix the bugs that have been occurring recently.

Trello card: https://trello.com/c/tL9InT1P/13-make-art-make-art-popup-keeps-popping-up-everytime-you-change-screen