RPi: Serve homepage for /challenge/.* routes
If the user launches a share from one of the challenge pack pages on
Kano OS and then clicks the "Back" button they were presented with a 404
page because Angular handles the page redirection instead of the server.
Resolve this by returning the homepage (main Angular app) from the
server when all routes of the form `/challenge/.*` are requested.