Browser: Disable the inspector
The inspector was disabled in 7566ce4 because it was preventing errors
created by 757edca from being correctly displayed to the user. This
unfortunately has the side-effect that it is possible to launch the
inspector with no way of being able to close it. Since the initial
commit, the underlying webkit has been updated and seems to not suffer
from the same problems so disable the inspector to prevent this.