[PC-168] Challenge tips disappear if challenge already complete (#437)
When a challenge is already complete and the browser has cached the solution and reapplies it 
when visiting the challenge, the tips get confused and show up blank. Resolve this by removing the offending `session.steps` which wasn't properly defined and thus allowing us to calculate the actual step that the user has reached.