[PC-160] Launching draw files by clicking them is unreliable (#443)
On Windows, it is possible to launch the app by clicking on one of the share files that you have downloaded separately. This uses a mechanism which Windows exposes into the JS and was being triggered at the same time as when we initialise the app, causing a race condition. Create an Angular service for handling this activation so that it can be called when we are ready for it rather than when it finishes.