Set up CI with Azure Pipelines
This sets up a CI/CD pipeline in Azure together with Store signing.
For the build pipeline to work, both the MakeArt_StoreKey.pfx
and MakeArt_StoreKey_Pass need to be configured in Azure DevOps.
For local development, create a temporary certificate and remember
to exclude that change from your final commits.