fix gutter from hiding text and cursor
when project length is >= 100 lines, the cursor in column position 0 will be hidden by the gutter being too wide and hiding the left-most part of the code editor. removing these lines in production seem to fix things, sadly as this is no longer an open-source project i can't verify this is the proper fix by building.